refactor(session): rename session.py to wiki.py and update module references
